fix: add separate models for VRF v1 and v2
